Added comments and sanity check

+++ src/main/java/org/apache/accumulo/testing/continuous/ContinuousVerify.java
@@ -197,13 +197,22 @@ public class ContinuousVerify extends Configured implements Tool {
 
       try {
         job.waitForCompletion(true);
+
+        // this will occur if verify is run without removing the output directory between runs
       } catch (FileAlreadyExistsException e) {
+        FileSystem fs = FileSystem.get(env.getHadoopConfiguration());
+
+        // if the exception occurred from something other than the output directory already
+        // existing, throw the exception
+        if (!fs.exists(outputPath)) {
+          throw e;
+        }
+
         String prompt = "\nOutput directory " + outputPath
             + " already exists. Do you want to delete it and re-run verify? [y/n] : ";
         String decision = System.console().readLine(prompt);
 
         if (decision.length() == 1 && decision.toLowerCase(Locale.ENGLISH).charAt(0) == 'y') {
-          FileSystem fs = FileSystem.get(env.getHadoopConfiguration());
           System.out.println("Deleting " + outputPath + " and retrying job.");
           fs.delete(outputPath, true);
           job.waitForCompletion(true);
